Refactored some code and folder structure to make it not as cluttured when adding Details sections for more markets.
