+++ src/client/pages/OfferNew/Introduction/Sidebar/DetailsModal/Details/components/SupportSection.tsx
