+++ src/client/pages/OfferNew/Introduction/Sidebar/DetailsModal/Details/components/index.tsx
@@ -1,0 +1,3 @@
+export * from './styledComponents'
+export * from './DetailInput'
+export * from './SupportSection'
