+++ src/client/pages/OfferNew/Introduction/Sidebar/DetailsModal/Details/index.tsx
@@ -1,0 +1,1 @@
+export * from './SwedishDetails'
