+++ src/client/pages/OfferNew/Introduction/Sidebar/DetailsModal/index.tsx
@@ -1,31 +1,18 @@
 import styled from '@emotion/styled'
 import { colorsV2, fonts } from '@hedviginsurance/brand'
-import { FieldArray, Form, Formik } from 'formik'
+import { Form, Formik } from 'formik'
 import React from 'react'
 import { Button } from 'components/buttons'
-import {
-  InputGroup,
-  InputGroupDeleteButton,
-  InputGroupRow,
-} from 'components/inputs'
 import { Modal, ModalProps } from 'components/ModalNew'
-import {
-  EditQuoteInput,
-  ExtraBuildingInput,
-  ExtraBuildingType,
-  useEditQuoteMutation,
-} from 'data/graphql'
+import { EditQuoteInput, useEditQuoteMutation } from 'data/graphql'
 import { OfferQuote } from 'pages/OfferNew/types'
 import { useTextKeys } from 'utils/textKeys'
-import { DetailInput } from './DetailInput'
-import { SupportSection } from './SupportSection'
 import {
   getFieldSchema,
   getInitialInputValues,
   getValidationSchema,
-  isApartmentFieldSchema,
-  isHouseFieldSchema,
 } from './utils'
+import { SwedishDetails } from './Details'
 
 const Container = styled.div`
   width: 100%;
@@ -56,70 +43,6 @@ const Headline = styled.div`
   @media (max-width: 600px) {
     font-size: 2rem;
     line-height: 3rem;
-  }
-`
-
-const Content = styled.div`
-  width: calc(100% + 2rem);
-  display: flex;
-  flex-direction: row;
-  margin: 0 -1rem;
-  margin-top: 2.5rem;
-
-  @media (max-width: 800px) {
-    width: 100%;
-    margin: 1.5rem 0 0 0;
-    flex-direction: column;
-  }
-`
-
-const ContentColumn = styled.div`
-  width: calc(50% - 2rem);
-  margin: 0 1rem;
-
-  @media (max-width: 800px) {
-    width: 100%;
-    margin: 0 0 2rem 0;
-
-    :last-child {
-      margin-bottom: 0;
-    }
-  }
-`
-
-const ContentColumnTitle = styled.div`
-  font-size: 1rem;
-  color: ${colorsV2.gray};
-  margin-bottom: 1rem;
-  padding-left: 0.5rem;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  min-height: 2rem;
-
-  @media (max-width: 800px) {
-    padding: 0;
-  }
-`
-
-const ContentColumnTitleButton = styled.button`
-  padding: 0.375rem 1rem;
-  font-size: 0.875rem;
-  color: ${colorsV2.gray};
-  border-radius: 24px;
-  border: 1px solid ${colorsV2.gray};
-  cursor: pointer;
-  transition: all 250ms;
-  background: none;
-
-  :focus {
-    outline: none;
-  }
-
-  :hover {
-    color: ${colorsV2.darkgray};
-    border-color: ${colorsV2.darkgray};
   }
 `
 
@@ -222,197 +145,12 @@ export const DetailsModal: React.FC<ModalProps & DetailsModalProps> = ({
           {(formikProps) => (
             <Form>
               <Headline>{textKeys.DETAILS_MODULE_HEADLINE()}</Headline>
-
-              {isApartmentFieldSchema(fieldSchema, offerQuote) && (
-                <Content>
-                  <ContentColumn>
-                    <InputGroup>
-                      <DetailInput
-                        field={fieldSchema.apartment.street}
-                        formikProps={formikProps}
-                        nameRoot="apartment"
-                        name="street"
-                      />
-
-                      <DetailInput
-                        field={fieldSchema.apartment.zipCode}
-                        formikProps={formikProps}
-                        nameRoot="apartment"
-                        name="zipCode"
-                      />
-
-                      <DetailInput
-                        field={fieldSchema.apartment.type}
-                        formikProps={formikProps}
-                        nameRoot="apartment"
-                        name="type"
-                      />
-
-                      <DetailInput
-                        field={fieldSchema.apartment.livingSpace}
-                        formikProps={formikProps}
-                        nameRoot="apartment"
-                        name="livingSpace"
-                      />
-
-                      <DetailInput
-                        field={fieldSchema.apartment.householdSize}
-                        formikProps={formikProps}
-                        nameRoot="apartment"
-                        name="householdSize"
-                      />
-                    </InputGroup>
-                  </ContentColumn>
-                  <ContentColumn>
-                    <SupportSection onButtonClick={() => Intercom('show')} />
-                  </ContentColumn>
-                </Content>
-              )}
-
-              {isHouseFieldSchema(fieldSchema, offerQuote) && (
-                <Content>
-                  <ContentColumn>
-                    <ContentColumnTitle>
-                      {textKeys.DETAILS_MODULE_TABLE_TITLE()}
-                    </ContentColumnTitle>
-                    <InputGroup>
-                      <DetailInput
-                        field={fieldSchema.house.street}
-                        formikProps={formikProps}
-                        nameRoot="house"
-                        name="street"
-                      />
-                      <DetailInput
-                        field={fieldSchema.house.zipCode}
-                        formikProps={formikProps}
-                        nameRoot="house"
-                        name="zipCode"
-                      />
-                      <InputGroupRow>
-                        <DetailInput
-                          field={fieldSchema.house.livingSpace}
-                          formikProps={formikProps}
-                          nameRoot="house"
-                          name="livingSpace"
-                        />
-                        <DetailInput
-                          field={fieldSchema.house.ancillarySpace}
-                          formikProps={formikProps}
-                          nameRoot="house"
-                          name="ancillarySpace"
-                        />
-                      </InputGroupRow>
-
-                      <InputGroupRow>
-                        <DetailInput
-                          field={fieldSchema.house.numberOfBathrooms}
-                          formikProps={formikProps}
-                          nameRoot="house"
-                          name="numberOfBathrooms"
-                        />
-                        <DetailInput
-                          field={fieldSchema.house.yearOfConstruction}
-                          formikProps={formikProps}
-                          nameRoot="house"
-                          name="yearOfConstruction"
-                        />
-                      </InputGroupRow>
-
-                      <DetailInput
-                        field={fieldSchema.house.householdSize}
-                        formikProps={formikProps}
-                        nameRoot="house"
-                        name="householdSize"
-                      />
-                      <DetailInput
-                        field={fieldSchema.house.isSubleted}
-                        formikProps={formikProps}
-                        nameRoot="house"
-                        name="isSubleted"
-                      />
-                    </InputGroup>
-                    <SupportSection onButtonClick={() => Intercom('show')} />
-                  </ContentColumn>
-                  <ContentColumn>
-                    <FieldArray
-                      name="house.extraBuildings"
-                      render={(arrayHelpers) => (
-                        <>
-                          <ContentColumnTitle>
-                            {textKeys.DETAILS_MODULE_EXTRABUILDINGS_TABLE_TITLE()}
-                            <ContentColumnTitleButton
-                              type="button"
-                              onClick={() => {
-                                const defaultExtraBuilding: ExtraBuildingInput = {
-                                  type: ExtraBuildingType.Garage,
-                                  area: 10,
-                                  hasWaterConnected: false,
-                                }
-                                arrayHelpers.insert(0, defaultExtraBuilding)
-                              }}
-                            >
-                              {textKeys.DETAILS_MODULE_EXTRABUILDINGS_TABLE_BUTTON()}
-                            </ContentColumnTitleButton>
-                          </ContentColumnTitle>
-
-                          {formikProps.values.house?.extraBuildings?.map(
-                            (_, index) => (
-                              <InputGroup key={index}>
-                                <DetailInput
-                                  field={fieldSchema.house.extraBuildings.type}
-                                  formikProps={formikProps}
-                                  nameRoot="house"
-                                  name={`extraBuildings.${index}.type`}
-                                />
-
-                                <DetailInput
-                                  field={fieldSchema.house.extraBuildings.area}
-                                  formikProps={formikProps}
-                                  nameRoot="house"
-                                  name={`extraBuildings.${index}.area`}
-                                />
-
-                                <DetailInput
-                                  field={
-                                    fieldSchema.house.extraBuildings
-                                      .hasWaterConnected
-                                  }
-                                  formikProps={formikProps}
-                                  nameRoot="house"
-                                  name={`extraBuildings.${index}.hasWaterConnected`}
-                                />
-
-                                <InputGroupDeleteButton
-                                  type="button"
-                                  onClick={() => {
-                                    const isLastItemLeft =
-                                      formikProps.values.house?.extraBuildings
-                                        ?.length === 1
-
-                                    arrayHelpers.remove(index)
-
-                                    if (isLastItemLeft) {
-                                      formikProps.setValues({
-                                        ...formikProps.values,
-                                        house: {
-                                          ...formikProps.values.house,
-                                          extraBuildings: [],
-                                        },
-                                      })
-                                    }
-                                  }}
-                                >
-                                  {textKeys.DETAILS_MODULE_EXTRABUILDINGS_TABLE_REMOVE_BUILDING_BUTTON()}
-                                </InputGroupDeleteButton>
-                              </InputGroup>
-                            ),
-                          )}
-                        </>
-                      )}
-                    />
-                  </ContentColumn>
-                </Content>
-              )}
+              <SwedishDetails
+                fieldSchema={fieldSchema}
+                formikProps={formikProps}
+                offerQuote={offerQuote}
+                textKeys={textKeys}
+              />
               <Footer>
                 <Button type="submit" disabled={isUpdating}>
                   {textKeys.DETAILS_MODULE_BUTTON()}
